Pre-cache the filter data since it takes a long time to compute everything
To accomplish this, we add a new tool, "importdictdata", which fills in the filter
and saves the state to disk. Then "indict" just loads the prefilled data, which
is much faster than recomputing it all from scratch.

Now that this is the case, change the desired false-positive rate from 50% to 5%
when constructing the filter.